Components folder structure following Atomic design renaming folders to elements, composites, sectionsœ
