+++ client/app/assets/styles/mixins.css
@@ -1,5 +1,4 @@
 @define-mixin guideStepListCheckbox {
-  background-image: url('../../assets/images/iconSquare@2x.png');
   background-repeat: no-repeat;
   background-size: 100%;
   background-position-y: 1px;
