+++ client/app/components/composites/SearchBar/SearchBar.css
