+++ client/app/components/composites/SearchBar/SearchBar.js
