+++ client/app/components/sections/OnboardingGuide/GuideBackToTodoLink.js
@@ -1,7 +1,7 @@
 import { PropTypes } from 'react';
 import { a } from 'r-dom';
 import css from './OnboardingGuide.css';
-import { t } from '../../utils/i18n';
+import { t } from '../../../utils/i18n';
 
 const GuideBackToTodoLink = (props) => {
