+++ client/app/components/sections/OnboardingGuide/OnboardingGuide.css
