+++ client/app/components/sections/OnboardingGuide/images/iconSquare@2x.png
