+++ client/app/components/sections/OnboardingGuide/images/iconSquareChecked@2x.png
