+++ client/app/components/sections/OnboardingGuide/images/step5_screenshot_paypal@2x.png
