+++ client/app/components/sections/OnboardingGuide/images/step7_screenshot_share@2x.png
