+++ client/app/components/sections/OnboardingTopBar/OnboardingTopBar.css
