+++ client/app/components/sections/Topbar/Topbar.story.js
@@ -1,5 +1,5 @@
 import { storiesOf } from '@kadira/storybook';
-import withProps from '../Styleguide/withProps';
+import withProps from '../../Styleguide/withProps';
 
 import Topbar from './Topbar';
