+++ client/app/startup/TopbarApp.js
@@ -1,6 +1,6 @@
 import r from 'r-dom';
 import { initialize as initializeI18n } from '../utils/i18n';
-import Topbar from '../components/Topbar/Topbar';
+import Topbar from '../components/sections/Topbar/Topbar';
 
 export default (props, railsContext) => {
   initializeI18n(railsContext.i18nLocale, railsContext.i18nDefaultLocale, process.env.NODE_ENV);
